- Fixed: Conversion of c_bind.cpp to FString was incomplete. - Fixed some functions that were declared as taking size_t's but defined as taking   unsigned ints. - Added a dummy object to delete sound environments on exit. - Fixed: FWarpTexture did not delete its Spans when destroyed. - Changed wadclusterinfos and wadlevelinfos arrays into TArrays. - Added the TypeInfo::AutoTypeInfoPtr for TypeInfo::m_RuntimeActors so they can   be automatically deleted. - Changed TypeInfo::m_Types into a TArray so it will be automatically deleted   on exit. - Fixed: TArray::Resize() did not deconstruct entries when shrinking the array. - Changed TArray::Push() so that it calls Grow() instead of duplicating the growth   calculations itself. - Calling TArray::Grow() for a small amount when the array is short should grow it   a bit more than it was doing.
SVN r76 (trunk)

+++ src/r_bsp.cpp
@@ -83,7 +83,7 @@ drawseg_t		*drawsegs;
 drawseg_t*		firstdrawseg;
 drawseg_t*		ds_p;
 
-unsigned int	FirstInterestingDrawseg;
+size_t			FirstInterestingDrawseg;
 TArray<size_t>	InterestingDrawsegs;
 
 fixed_t			WallTX1, WallTX2;	// x coords at left, right of wall in view space
